working on payables final

+++ database/migrations/2022_09_18_172700_create_inventories_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('store_id');
             $table->unsignedBigInteger('item_id');
             $table->unsignedBigInteger('investor_id');
-            $table->float('unit_cost');
+            $table->double('unit_cost');
             $table->unsignedBigInteger('quantity');
             $table->timestamps();
         });
